requirements.txt
docx2txt==0.8
gradio==3.39.0
openai==0.27.8
pdf2docx==0.5.6
PyYAML==6.0.1
